Rename AS/IXP page to networks
