+++ src/views/charts/tables/HegemonyAlarmsTable.vue
@@ -19,7 +19,7 @@
           <q-toggle v-model="props.expand" />
         </q-td>
         <q-td key="originasn" align>
-          <a @click="newWindow({name : 'as_and_ixp', params:{asn: $options.filters.ihr_NumberToAsOrIxp(props.row.originasn)}})" href="javascript:void(0)">
+          <a @click="newWindow({name : 'networks', params:{asn: $options.filters.ihr_NumberToAsOrIxp(props.row.originasn)}})" href="javascript:void(0)">
             {{$options.filters.ihr_NumberToAsOrIxp(props.row.originasn)}}
           </a>
         </q-td>
